Fix the UV Index data types (they're not integers)

+++ database/upgrade_v3.sql
@@ -85,10 +85,10 @@ create table davis_sample (
   solar_radiation float,
   wind_sample_count int,
   gust_wind_direction float,
-  average_uv_index int,
+  average_uv_index numeric(2,1),
   evapotranspiration float,
   high_solar_radiation float,
-  high_uv_index int,
+  high_uv_index numeric(2,1),
   forecast_rule_id int
 
 -- These columns are not currently stored as I've no way of testing them with
@@ -347,7 +347,7 @@ create table davis_live_data (
   console_battery_voltage float,
   forecast_icon int,
   forecast_rule_id int,
-  uv_index int,
+  uv_index numeric(2,1),
   solar_radiation int
 );
 comment on table davis_live_data is 'Additional live data available from Davis-compatible hardware';
@@ -361,8 +361,9 @@ comment on column davis_live_data.forecast_icon is 'Forecast icon';
 comment on column davis_live_data.forecast_rule_id is 'Current forecast rule. See davis_forecast_rule table for values';
 comment on column davis_live_data.uv_index is 'Latest UV index reading';
 comment on column davis_live_data.solar_radiation is 'Latest solar radiation reading in watt/meter squared';
+
 ----------------------------------------------------------------------
--- CONSTRAINTS ----------------------------------------------------------
+-- CONSTRAINTS -------------------------------------------------------
 ----------------------------------------------------------------------
 
 -- Don't allow duplicate timestamps within a station.
